add nlp module base class
