+++ paddlehub/__init__.py
@@ -63,5 +63,3 @@ from .finetune.strategy import ULMFiTStrategy
 from .finetune.strategy import CombinedStrategy
 
 from .autofinetune.evaluator import report_final_result
-
-from .module.nlp_module import BERTModule
